requirements.txt
pandas
numpy
pyyaml
scikit-learn
yfinance
xgboost
pyarrow
requests
feedparser
nltk
deep-translator

scikit-optimize
